requirements.txt
django>=4.2.2
black[d]>=23.3.0
pika>=1.3.2
djangorestframework>=3.14.0